separated responsibilities, simplified a little, added tests for imports.
